score 100/100 for collinear

+++ Hw3_Notwork/Point.java
@@ -13,6 +13,7 @@ import edu.princeton.cs.algs4.StdDraw;
 import java.util.Comparator;
 
 public class Point implements Comparable<Point> {
+
     private final int x;     // x-coordinate of this point
     private final int y;     // y-coordinate of this point
 
@@ -68,7 +69,7 @@ public class Point implements Comparable<Point> {
             return Double.POSITIVE_INFINITY;
         }
         else {
-            return (double) (that.y - this.y) / (that.x - this.x);
+            return (that.y - this.y) / (that.x - this.x);
         }
     }
